+++ src/constants.ts
@@ -1,4 +1,4 @@
-import { parseUnits } from "ethers/utils"
+import { utils } from "ethers"
 
 export const ipfsGatewayListUri = "https://ipfs.github.io/public-gateway-checker/gateways.json"
 
@@ -15,13 +15,13 @@ export const vocdoniSokolEntityId = "0xe00f56aac24b5b78b94c38f511631b8b09a5c5a9e
 
 export const XDAI_CHAIN_ID = 100
 export const XDAI_PROVIDER_URI = "https://dai.poa.network"
-export const XDAI_GAS_PRICE = parseUnits("1", "gwei")
+export const XDAI_GAS_PRICE = utils.parseUnits("1", "gwei")
 export const XDAI_ENS_REGISTRY_ADDRESS = "0x00cEBf9E1E81D3CC17fbA0a49306EBA77a8F26cD"
 export const XDAI_TEST_ENS_REGISTRY_ADDRESS = "0x9e638E90c8CdFaC1297EF261859E25c9d8438F1a"
 
 export const SOKOL_CHAIN_ID = 77
 export const SOKOL_PROVIDER_URI = "https://sokol.poa.network"
-export const SOKOL_GAS_PRICE = parseUnits("1", "gwei")
+export const SOKOL_GAS_PRICE = utils.parseUnits("1", "gwei")
 export const SOKOL_ENS_REGISTRY_ADDRESS = "0x43541c49308bF2956d3893836F5AF866fd78A295"
 
 // export const SIGNATURE_TIMESTAMP_TOLERANCE = 60 * 8 // +/- 8 minutes